+++ yash-env/CHANGELOG.md
@@ -31,6 +31,8 @@ A _private dependency_ is used internally and not visible to downstream users.
     - `Read`: Declares the `read` method for reading from file descriptors.
     - `Seek`: Declares the `lseek` method for seeking within file
       descriptors.
+    - `Umask`: Declares the `umask` method for setting the file mode
+      creation mask.
     - `Write`: Declares the `write` method for writing to file
       descriptors.
 - Implementations of these traits are provided for those types that implement